+++ src/Services/Posts/InstaConnect.Posts.Infrastructure/Features/Posts/Abstractions/IPostCollectionFactory.cs
@@ -4,5 +4,5 @@ using InstaConnect.Posts.Domain.Features.Posts.Models.Responses;
 namespace InstaConnect.Posts.Infrastructure.Features.Posts.Abstractions;
 internal interface IPostCollectionFactory
 {
-    PostCollection Create(ICollection<Post> posts, int totalCount, PostPaginationQuery pagination);
+    PostCollection Create(ICollection<Post> posts, long totalCount, PostPaginationQuery pagination);
 }
